refactor : AuthenticationInterceptor 추상화
