+++ src/main/java/nextstep/subway/auth/ui/common/AuthenticationInterceptor.java
@@ -1,0 +1,49 @@
+package nextstep.subway.auth.ui.common;
+
+import nextstep.subway.auth.domain.Authentication;
+import nextstep.subway.auth.domain.AuthenticationToken;
+import nextstep.subway.member.application.CustomUserDetailsService;
+import nextstep.subway.member.domain.LoginMember;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.Optional;
+
+public abstract class AuthenticationInterceptor implements HandlerInterceptor {
+    private CustomUserDetailsService userDetailsService;
+    private AuthenticationConverter converter;
+
+    public AuthenticationInterceptor(CustomUserDetailsService userDetailsService, AuthenticationConverter converter) {
+        this.userDetailsService = userDetailsService;
+        this.converter = converter;
+    }
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws IOException {
+        AuthenticationToken token = converter.convert(request);
+        Authentication authentication = authenticate(token);
+        afterAuthentication(request, response, authentication);
+        return false;
+    }
+
+    public Authentication authenticate(AuthenticationToken token) {
+        String principal = token.getPrincipal();
+        LoginMember userDetails = userDetailsService.loadUserByUsername(principal);
+        checkAuthentication(userDetails, token);
+        return Optional.of(new Authentication(userDetails)).orElseThrow(RuntimeException::new);
+    }
+
+    private void checkAuthentication(LoginMember userDetails, AuthenticationToken token) {
+        if (userDetails == null) {
+            throw new RuntimeException();
+        }
+
+        if (!userDetails.checkPassword(token.getCredentials())) {
+            throw new RuntimeException();
+        }
+    }
+
+    public abstract void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException;
+}
